+++ instrumentation/grpc-1.6/build.gradle.kts
@@ -62,6 +62,7 @@ val grpcVersion = "1.6.0"
 dependencies {
     api("io.opentelemetry.javaagent.instrumentation:opentelemetry-javaagent-grpc-1.6:${versions["opentelemetry_java_agent"]}")
     api("io.opentelemetry.instrumentation:opentelemetry-grpc-1.6:${versions["opentelemetry_java_agent"]}")
+    implementation(project(":instrumentation:grpc-common"))
 
     compileOnly("io.grpc:grpc-core:${grpcVersion}")
     compileOnly("io.grpc:grpc-protobuf:${grpcVersion}")
