+++ instrumentation/grpc-common/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    `java-library`
+}
+
+dependencies {
+    compileOnly("io.grpc:grpc-core:1.6.0")
+}
